Initialise dependencies
We have a rough idea what the application needs at the moment. We're
adding some of the dependencies to save ourself some grief developing
them ourselves.

`volume-go` is a package that manages the volume cross operating system.
Meaning we can send a signal to the application in order to silence,
boost, mute and unmute our machine. Pretty neat.

`beeep` is the library that provides notifications cross platform. We
believe it will come in useful to copy the notification from one device
to the other making all the machines connected to the ecosystem aware
what's going on.

`logrus` is simply a library that provides logging. We think it is
useful when it comes to configuration of when and how the logs are being
presented.

`clipboard` provides cross platform clipboard management. It requires
`xsel` or `xclip` on GNU/Linux systems. It would come in handy if we'd
like to copy a piece of information from one device to another.